refactor: Refactor code and directories
