+++ src/main/java/com/project/dealer_api/domain/dealer/Dealer.java
@@ -1,4 +1,4 @@
-package com.project.dealer_api.models;
+package com.project.dealer_api.domain.dealer;
 
 import javax.persistence.*;
