+++ src/main/java/com/project/dealer_api/infra/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.project.dealer_api.config;
+package com.project.dealer_api.infra.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
