+++ src/main/java/com/project/dealer_api/infra/security/TokenService.java
@@ -4,7 +4,7 @@ import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.exceptions.JWTCreationException;
 import com.auth0.jwt.exceptions.JWTVerificationException;
-import com.project.dealer_api.models.Users;
+import com.project.dealer_api.domain.user.Users;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
