+++ src/main/java/com/project/dealer_api/repository/DealerRepository.java
@@ -1,6 +1,6 @@
 package com.project.dealer_api.repository;
 
-import com.project.dealer_api.models.Dealer;
+import com.project.dealer_api.domain.dealer.Dealer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
