+++ src/main/java/com/project/dealer_api/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.project.dealer_api.repository;
 
-import com.project.dealer_api.models.Users;
+import com.project.dealer_api.domain.user.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.security.core.userdetails.UserDetails;
